streamlit-app/requirements.txt
streamlit==1.40.2
supabase==2.14.0
pandas==2.2.3
plotly==5.24.1
python-dotenv==1.0.1
httpx==0.27.2
